requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
apscheduler>=3.10.0
line-bot-sdk>=3.14.0
python-dotenv>=1.0.0
pytz>=2023.3

# 自動站內信功能依賴
PyPtt>=1.0.0
google-generativeai>=0.3.0
